錄影設備測試OK

+++ recording/Task_check_frsip.php
@@ -5,14 +5,16 @@
 	
 	global $g_create_meeting_apiurl, $g_prod_meeting_apiurl;
 	
-	wtask_log("Task_check_meetingroom entry <-");
+	$mainurl = $g_create_meeting_apiurl;
 	set_time_limit(0);
 	try
 	{
+		$remote_ip4filename = get_remote_ip_underline();
+		wtask_log("Task_check_frsip", $remote_ip4filename, "Task_check_meetingroom entry <-");
 
-		if(file_exists("/tmp/check_frsip.pid")==true)//還在跑
+		if (file_exists("/tmp/check_frsip.pid") == true) // 還在跑
 		{
-			if(strtotime(date("Y-m-d H:i:s")) - filemtime("/tmp/check_frsip.pid")> (3*60*60))//超過3小時
+			if(strtotime(date("Y-m-d H:i:s")) - filemtime("/tmp/check_frsip.pid") > 3*60*60) // 超過3小時
 			{
 				// 可能不正常離開
 			}
@@ -20,51 +22,51 @@
 			{
 				$msg = strtotime(date("Y-m-d H:i:s"))." - ".filemtime("/tmp/check_frsip.pid");
 				echo $msg."\r\n";
-				wtask_log($msg."\r\n".$g_exit_symbol."Task_check_frsip exit ->"."\r\n");
+				wtask_log("Task_check_frsip", $remote_ip4filename, $msg."\r\n".$g_exit_symbol."Task_check_frsip exit ->"."\r\n");
 				return;
 			}
+			touch("/tmp/check_frsip.pid");
 		}
-		touch("/tmp/check_frsip.pid");
-
+		
+		// connect mysql
 		$link = mysqli_connect($host, $user, $passwd, $database);
-		mysqli_query($link,"SET NAMES 'utf8'");
-
-		$mainurl = $g_create_meeting_apiurl
-		$url = $mainurl."post/api/token/request";
-
+		$data = result_connect_error ($link);
+		if ($data["status"] == "false")
+		{
+			wtask_log("Task_check_frsip", $remote_ip4filename, "[Task_check_frsip] ".get_error_symbol($data["code"])." query result :".$data["code"]." ".$data["responseMessage"]."\r\n".$g_exit_symbol."send otp exit ->"."\r\n");
+			return;
+		}
+		mysqli_query($link, "SET NAMES 'utf8'");
+		
 		//1. GET Token
-		$data 				= array();
-		//$data["username"]	= "administrator";
-		$data["username"]	= "administrator";
-		$hash 				= md5("CheFR63r");
-		//$hash 			= md5("sT7m");
-		$data["data"]		= md5($hash."@deltapath");
-		//echo md5($hash."@deltapath");
-		$out = CallAPI4OptMeeting("POST", $url, $data);
-		//echo $out;
+		$out = get_meeting_token("Task_check_frsip", $g_create_meeting_apiurl, $remote_ip4filename, $g_meeting_uid, $g_meeting_pwd);
+		if (strpos($out, "\"success\"") == false) return;
+		
 		$ret = json_decode($out, true);
 		if($ret['success'] == true)
+		{
+			echo "get token succeed\r\n";
 			$token = $ret['token'];
+		}
 		else
 		{
 			//第二次機會
 			if(_ENV == "PROD")
 			{
-				$mainurl 			= $g_prod_meeting_apiurl;
-				$url 				= $mainurl."post/api/token/request";
-				$data 				= array();
-				$data["username"]	= "administrator";
-				$hash 				= md5("CheFR63r");
-				$data["data"]		= md5($hash."@deltapath");
-				$out = CallAPI4OptMeeting("POST", $url, $data);
-				//echo $out;
+				$mainurl = $g_prod_meeting_apiurl;
+				$out = get_meeting_token("Task_check_frsip", $g_prod_meeting_apiurl, $remote_ip4filename, $g_meeting_uid, $g_meeting_pwd);
+				if (strpos($out, "\"success\"") == false) return;
+				
 				$ret = json_decode($out, true);
 				if($ret['success'] == true)
+				{
+					echo "get prod token succeed\r\n";
 					$token = $ret['token'];
+				}
 				else
 				{
 					echo "error";//error;
-					unlink("/tmp/check_frsip.pid");
+					if (file_exists("/tmp/check_frsip.pid") == true) unlink("/tmp/check_frsip.pid");
 					//寫入資料庫, server error
 					$sql = "update vmrule set frsipstatus = 1  where 1";
 					$result = mysqli_query($link, $sql);		
@@ -72,23 +74,26 @@
 				}
 			}
 		}
-
-		$header = array('X-frSIP-API-Token:'.$token);
-
-		$url 				= $mainurl."get/systemstatus/serverstatus/2";
-		$data 				= array();
-		$data["serverId"]	= "2";
-		$out = CallAPI4OptMeeting("GET", $url, $data, $header);
-		echo $out;
+		
+		$header 					= array('X-frSIP-API-Token:'.$token);
+		$url 						= $mainurl."get/systemstatus/serverstatus/2";
+		$data_input02 				= array();
+		$data_input02["serverId"]	= "2";
+		wtask_log("Task_check_frsip", $remote_ip4filename, "api url :".$url);
+		wtask_log("Task_check_frsip", $remote_ip4filename, "serverId :".$data_input02["serverId"]);
+		$out = CallAPI4OptMeeting("GET", $url, $data_input02, $header);
+		wtask_log("Task_check_frsip", $remote_ip4filename, "query serverstatus api result :".$out);
+		echo "query serverstatus api result :".$out."\r\n";
 		if (strlen($out) <= 0)
 		{
 			//寫入資料庫, server error
-			unlink("/tmp/check_frsip.pid");
+			if (file_exists("/tmp/check_frsip.pid") == true) unlink("/tmp/check_frsip.pid");
 			$sql = "update vmrule set frsipstatus = 2  where 1";
 			$result = mysqli_query($link, $sql);		
 			return;
 		}
-
+		if (strpos($out, "\"disk\"") == false) return;
+		
 		$ret = json_decode($out, true);
 		if (strlen($ret['disk']) > 0)
 		{
@@ -99,7 +104,7 @@
 			if(strstr($diskusage, "100%") && strstr($diskusage, "99%"))
 			{
 				//error 
-				unlink("/tmp/check_frsip.pid");
+				if (file_exists("/tmp/check_frsip.pid") == true) unlink("/tmp/check_frsip.pid");
 				$sql = "update vmrule set frsipstatus = 3 where 1";
 				$result = mysqli_query($link, $sql);			
 				return;
@@ -107,7 +112,7 @@
 		}
 		if ($ret['status'] == "Offline")
 		{
-			unlink("/tmp/check_frsip.pid");
+			if (file_exists("/tmp/check_frsip.pid") == true) unlink("/tmp/check_frsip.pid");
 			$sql = "update vmrule set frsipstatus = 5 where 1";
 			$result = mysqli_query($link, $sql);			
 			return;	
@@ -121,7 +126,7 @@
 		if (strlen($out) <= 0)
 		{
 			//寫入資料庫, server error
-			unlink("/tmp/check_frsip.pid");
+			if (file_exists("/tmp/check_frsip.pid") == true) unlink("/tmp/check_frsip.pid");
 			$sql = "update vmrule set frsipstatus = 4  where 1";
 			$result = mysqli_query($link, $sql);		
 			return;
@@ -142,7 +147,7 @@
 
 		if ($ret['status'] == "Offline")
 		{
-			unlink("/tmp/check_frsip.pid");
+			if (file_exists("/tmp/check_frsip.pid") == true) unlink("/tmp/check_frsip.pid");
 			$sql = "update vmrule set frsipstatus = 6 where 1";
 			$result = mysqli_query($link, $sql);			
 			return;	
@@ -152,7 +157,7 @@
 		$sql = "update vmrule set frsipstatus = 0  where 1";
 		$result = mysqli_query($link, $sql);
 
-		unlink("/tmp/check_frsip.pid");
+		if (file_exists("/tmp/check_frsip.pid") == true) unlink("/tmp/check_frsip.pid");
 
 		$data = array();
 		// expire token 
@@ -163,12 +168,12 @@
 	}
 	catch (Exception $e)
 	{
-		wtask_log_Exception("Exception error :".$e->getMessage());
+		wtask_log_Exception("Task_check_frsip", $remote_ip4filename, "Exception error :".$e->getMessage());
 		echo "error";
 	}
 	finally
 	{
-		wtask_log("finally procedure");
+		wtask_log("Task_check_frsip", $remote_ip4filename, "finally procedure");
 		try
 		{
 			if ($link != null)
@@ -179,8 +184,8 @@
 		}
 		catch (Exception $e)
 		{
-			wtask_log_Exception("Exception error: disconnect! error :".$e->getMessage());
+			wtask_log_Exception("Task_check_frsip", $remote_ip4filename, "Exception error: disconnect! error :".$e->getMessage());
 		}
-		wtask_log("finally complete"."\r\n".$g_exit_symbol."Task_check_frsip exit ->"."\r\n");
+		wtask_log("Task_check_frsip", $remote_ip4filename, "finally complete"."\r\n".$g_exit_symbol."Task_check_frsip exit ->"."\r\n");
 	}
 ?>
